PythonDemo/requirements.txt
Flask==3.0.0
requests==2.31.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
flask-cors==4.0.0
